Fixed pager / checkboxes  / final issue creation msg

+++ src/main/java/ru/mail/jira/plugins/myteam/rulesengine/core/Pager.java
@@ -35,7 +35,7 @@ public class Pager {
   }
 
   public boolean hasNext() {
-    return Math.ceil(total / pageSize) - page > 1;
+    return Math.ceil((float) total / pageSize) - page > 1;
   }
 
   public boolean hasPrev() {
